+++ Changes.md
@@ -6,6 +6,11 @@ Fixes
 
 - Spreadsheet : Fixed attempts to edit non-editable plugs when double-clicking on a boolean cell.
 
+API
+---
+
+- SpreadsheetUI : Added `addColumnButtonMenuSignal()`, which can be used to customise the column creation menu for spreadsheets.
+
 0.61.5.0 (relative to 0.61.4.0)
 ========